removed commented code and print statement
